requirements_full.txt
# Core Development Tools
black==25.1.0
flake8==7.3.0
isort==6.0.1
mypy==1.17.1
pytest==8.4.1

# Web Framework (for Vercel API)
Flask==3.1.0
python-dotenv==1.0.1
psycopg[binary]>=3.2.0

# Trading & Finance
ccxt==4.5.0
python-okx==0.4.0
requests==2.32.4
pandas==2.3.1
numpy==2.3.2
matplotlib==3.10.5

# Real-time & Performance
websockets==15.0.1
redis==6.4.0
structlog==25.4.0

# Visualization
plotly==6.3.0
mplfinance==0.12.10b0

# Data Processing
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Networking & Security
aiohttp==3.12.15
cryptography==45.0.6
certifi==2025.8.3
urllib3==1.26.12

# ML (optional)
